Scoring, add: * Custom phrase query. Favour phrases near beginning of article * Cached rank field source, provides rank info for docid * Raw score transformations (e.g. LogTransform) * Unit tests for ranking Misc: * Related mapping is built from links appearing in same context * Introduce token gaps in Tokenizer, move initial templates to after   the introduction text, put gaps between paragraphs, sections, ... * Fix bug with not closing the streams after an socket.accept() error * RelatedMap for related-article score propagation, experimental,   probably will be reverted
git-svn-id: file:///home/git-svn-mirror/pushmi-svn-mirror/branches/lucene-search-2.1@26576 dd0e9695-b195-4be7-bd10-2dea1a65a6b6

+++ src/org/wikimedia/lsearch/related/Related.java
@@ -54,7 +54,7 @@ public class Related {
 	}
 	
 	public String serialize(){
-		return score+" "+relates;
+		return (float)score+" "+relates;
 	}
 	
 	public String getRelates() {
